aidl: fingerprint: Introduce SecUdfpsHelper
SecUdfpsHelper is a Fingerprint HAL addition that helps with UDFPS related operations on Samsung devices.

It currently has two different purposes:

* Providing a control over touchscreen.
* Providing a way of monitoring HBM state on Qualcomm Samsung devices,
with the help of newly introduced DRM property.

1- Touchscreen control
-------------------------------------------------------------------
Touchscreen control offers two different ways to do this

* Default method: Using set/get functions to write commands to the touchscreen cmd node.
* SysInput method: Using Samsung SysInput service.

Default method provides a more legacy way of handling, while SysInput method can be
useful for foldable devices, i.e. devices with multiple input hardwares. The SysInput HAL
provides communication with each input hardware (Multi-TSP, Keyboard etc.) so that we don't have to define each path
individually.

Note that the SysInput method is completely optional and the maintainer is not required to provide the HAL for their devices.

2- HBM monitoring (Qualcomm only)
---------------------------------------------------------------------
HBM monitoring requires the newly introduced DRM feature and 2 new sysfs interfaces to work properly.

HBM monitoring will register an inotify instance to the HBM status path and start listening.
When the status path is changed, it will read the file and if HBM is enabled, it will write the HBM brightness value to mask_brightness sysfs.

Since each device needs a specific brightness value due to different screens, maintainers can expose the required brightness value via sysfs. Fingerprint HAL will read and use this value to write the HBM brightness.

If sysfs is not exposed, Fingerprint HAL will use the default value of 319.

Note that this monitoring logic was written with Framework dimming in mind.
Kernel dimming implementations may not work or may require an adaptation.

Change-Id: I17b54744d7ce2cdcd40b175bafec06eb48894db5